+++ HISTORY.md
@@ -1,12 +1,13 @@
 # History
 
-## (Unreleased)
+## (Forthcoming)
 
 ### Changelog
 
 #### API
 
 - Add `DDS-CNF` density method to `DensityArchive` ({pr}`691`)
+- Add `NSLCRanker` for Novelty Search with Local Competition ({pr}`690`)
 
 ## 0.10.0